+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - Open AI configuration setting `organizationId` is not used when verifying configuration.
 - Gemini configuration settings `projectId` and `location` are not used when verifying configuration.
+- Notification about common branch is shown after the prompt dialog is closed.
 
 ## [2.4.1] - 2024-09-19